+++ src/Constants.ts
@@ -1,4 +1,5 @@
 export const DEFAULT_OFFSET = 50
+export const DDP_LOG_RETENTION_LIMIT = 5000
 
 export const DEVELOPMENT = import.meta.env.MODE === 'development'