+++ src/Stores/Panel/MinimongoStore/index.ts
@@ -129,8 +129,6 @@ export class MinimongoStore {
   ): IDocumentWrapper {
     const _string = JSONUtils.stringify(document)
 
-    console.log({ collectionName })
-
     return {
       collectionName,
       document,